menambah logo dan membuat level admin

+++ application/controllers/Admin.php
@@ -12,6 +12,9 @@ class Admin extends CI_Controller
 
     public function index()
     {
+        if ($this->session->userdata('level') != 'administrator') {
+            redirect('auth', 'refresh'); // Redirect jika bukan administrator
+        }
         // Tangkap data
         $data['title'] = 'Admin';
         $data['admin'] = $this->M_admin->GetAllAdmin();
@@ -26,12 +29,17 @@ class Admin extends CI_Controller
 
     public function create()
     {
+        // Pemeriksaan apakah admin atau administrator
+        if ($this->session->userdata('level') != 'administrator') {
+            redirect('auth', 'refresh'); // Redirect jika bukan administrator
+        }
         $data['title'] = 'Tambah Data Admin';
         $data['namaAdmin'] = $this->db->get_where('admin', ['id_admin' => $this->session->userdata('id_admin')])->row_array();
 
         $this->form_validation->set_rules('id_admin', 'Username', 'required');
         $this->form_validation->set_rules('nama_admin', 'Nama', 'required');
         $this->form_validation->set_rules('pswd_admin', 'Password', 'required');
+        $this->form_validation->set_rules('level', 'Level', 'required');
 
         if ($this->form_validation->run() == FALSE) {
             $this->load->view('backend/templates/header', $data);
@@ -48,6 +56,10 @@ class Admin extends CI_Controller
 
     public function edit($id_admin)
     {
+        // Pemeriksaan apakah admin atau administrator
+        if ($this->session->userdata('level') != 'administrator') {
+            redirect('auth', 'refresh'); // Redirect jika bukan administrator
+        }
         $data['title'] = 'Edit Data Admin';
         $data['admin'] = $this->M_admin->getAdminById($id_admin);
         $data['namaAdmin'] = $this->db->get_where('admin', ['id_admin' => $this->session->userdata('id_admin')])->row_array();
@@ -55,6 +67,7 @@ class Admin extends CI_Controller
         $this->form_validation->set_rules('id_admin', 'Username', 'required');
         $this->form_validation->set_rules('nama_admin', 'Nama', 'required');
         $this->form_validation->set_rules('pswd_admin', 'Password', 'required');
+        $this->form_validation->set_rules('level', 'Level', 'required');
 
         if ($this->form_validation->run() == FALSE) {
             $this->load->view('backend/templates/header', $data);
@@ -66,7 +79,8 @@ class Admin extends CI_Controller
             $adminData = array(
                 'id_admin' => $this->input->post('id_admin'),
                 'nama_admin' => $this->input->post('nama_admin'),
-                'pswd_admin' => md5($this->input->post('pswd_admin'))
+                'pswd_admin' => md5($this->input->post('pswd_admin')),
+                'level' => $this->input->post('level')
             );
 
             $this->M_admin->edit($id_admin, $adminData);
@@ -83,6 +97,11 @@ class Admin extends CI_Controller
 
     public function delete($id_admin)
     {
+        // Pemeriksaan apakah admin atau administrator
+        if ($this->session->userdata('level') != 'administrator') {
+            redirect('auth', 'refresh'); // Redirect jika bukan administrator
+        }
+
         $this->M_admin->delete($id_admin);
         $this->session->set_flashdata('flash', 'Dihapus');
         redirect('admin');
